Return only the project output in managed metaproj
Fixes #396.

There are two ways to define project-to-project references in MSBuild, the
"MSBuild way" with `ProjectReference` items that turn into `MSBuild` task
invocations, and the "old Visual Studio way", which puts information about
dependencies into the `.sln` file.

When the latter method is used and MSBuild is instructed to build the
`.sln` file, it creates in-memory MSBuild projects that use the `MSBuild`
task to express the dependencies between projects.

Prior to this commit, the `.metaproj` for a managed product had a `Build`
target that returned an item group consisting of:
 * The output of the project that the `.metaproj` wrapped, AND
 * The outputs of all of the projects that were ordered before this
   project in the solution build order.

Since those projects in turn returned the outputs of all of *their*
dependencies, this list could grow to enormous size in a medium-to-large
solution. A customer reported consistent out-of-memory crashes when
building a solution of around 200 projects.

This commit removes the second category from the returned output of the
metaproject. This restores parity with metaprojects produced for "Web
Projects", with projects that are built directly without generating a
metaproject (as when they define no solution build order), and with
building projects directly without building a containing solution.